Move the get_mcm.py and get_py.py scripts to the python folder and make both of them importable modules. Add a new script to run get_py.py multiple times using different dictionaries and output folders with all other options being passed through.
